chore: Relocate PersistentXssUtils
- Add maintenance note to changelogs.
- Move PersistentXssUtils to commonlib, renaming it to SourceSinkUtils.
- Update ascanrules dependant files to use the renamed class.
- Update ascanrules build file to depend on commonlib 1.9.0+.
- Also addressed a few SAST (SonarLint) items in SourceSinkUtils.
