+++ addOns/ascanrules/ascanrules.gradle.kts
@@ -14,7 +14,7 @@ zapAddOn {
         dependencies {
             addOns {
                 register("commonlib") {
-                    version.set(">= 1.6.0 & < 2.0.0")
+                    version.set(">= 1.9.0 & < 2.0.0")
                 }
             }
         }
